+++ inc/ticketsolveddate.class.php
@@ -3,7 +3,7 @@
 /**
  * -------------------------------------------------------------------------
  * YAGP plugin for GLPI
- * Copyright (C) 2019-2024 by the TICgal Team.
+ * Copyright (C) 2019-2025 by the TICgal Team.
  * https://tic.gal/en/project/yagp-yet-another-glpi-plugin/
  * -------------------------------------------------------------------------
  * LICENSE
@@ -18,24 +18,22 @@
  * GNU General Public License for more details.
  * You should have received a copy of the GNU General Public License
  * along with YAGP. If not, see <http://www.gnu.org/licenses/>.
- * --------------------------------------------------------------------------
- * @package   YAGP
- * @author    the TICgal team
- * @copyright Copyright (c) 2019-2024 TICgal team
+ * -------------------------------------------------------------------------
+ * @package   yagp
+ * @author    the TICGAL team
+ * @copyright Copyright (c) 2025 TICGAL team
  * @license   AGPL License 3.0 or (at your option) any later version
  *            http://www.gnu.org/licenses/agpl-3.0-standalone.html
- * @link      https://tic.gal/en/project/yagp-yet-another-glpi-plugin/
+ * @link      https://www.tic.gal
  * @since     2019
- * ----------------------------------------------------------------------
+ * -------------------------------------------------------------------------
  */
 
+// phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNamespace
 class PluginYagpTicketsolveddate extends CommonDBTM
 {
     /**
-     * getTypeName
-     *
-     * @param  mixed $nb
-     * @return string
+     * {@inheritdoc}
      */
     public static function getTypeName($nb = 0): string
     {
@@ -45,26 +43,34 @@ class PluginYagpTicketsolveddate extends CommonDBTM
     /**
      * cronInfo
      *
-     * @param  mixed $name
+     * @param  string $name
      * @return array
      */
-    public static function cronInfo($name): array
+    public static function cronInfo(string $name): array
     {
         switch ($name) {
             case 'changeDate':
                 return [
                     'description'   => __('Change date', 'yagp'),
-                    'parameter'     => __('Number of tickets', 'yagp')
+                    'parameter'     => __('Number of tickets', 'yagp'),
                 ];
         }
+
         return [];
     }
 
-    public static function cronChangeDate($task)
+    /**
+     * @param CronTask $task
+     *
+     * @return int
+     */
+    public static function cronChangeDate(CronTask $task): int
     {
+        /** @var \DBmysql $DB */
         global $DB;
 
         $config = PluginYagpConfig::getInstance();
+        $tot = 0;
         if ($config->fields['ticketsolveddate']) {
             $message = "";
             $ticket = new Ticket();
@@ -74,81 +80,27 @@ class PluginYagpTicketsolveddate extends CommonDBTM
                 $limit = "";
             }
 
-           /*$sub_task=new QuerySubQuery([
-            'SELECT'=>[
-               'tickets_id',
-               'MAX'=>'end AS last_task_end',
-            ],
-            'FROM'=>'glpi_tickettasks',
-            'WHERE'=>[
-               [
-                  'NOT' => ['end' => null],
-               ],
-            ],
-            'GROUPBY'=>'tickets_id',
-            'AS task'
-           ]);
-           $sub_taskstart=new QuerySubQuery([
-            'SELECT'=>[
-               'tickets_id',
-               'MIN'=>'begin AS first_task_begin',
-            ],
-            'FROM'=>'glpi_tickettasks',
-            'WHERE'=>[
-               [
-                  'NOT' => ['begin' => null],
-               ],
-            ],
-            'GROUPBY'=>'tickets_id',
-            'AS taskstart'
-           ]);
-           $query=[
-            'SELECT'=>[
-               'id',
-               'date',
-               'solvedate',
-            ],
-            'FROM'=>'glpi_tickets',
-            'INNER JOIN'=>[
-               $sub_task=>[
-                  'FKEY'=>[
-                     'glpi_tickets'=>'id',
-                     'task'=>'tickets_id',
-                  ]
-               ],
-               $sub_taskstart=>[
-                  'FKEY'=>[
-                     'glpi_tickets'=>'id',
-                     'taskstart'=>'tickets_id',
-                  ]
-               ]
-            ],
-            'WHERE'=>[
-               'status'=>['>=','5'],
-               'is_deleted'=>0,
-               'solvedate'=>['!=','last_task_end']
-            ]
-           ];*/
+            // TODO: transform into array for iterator
             $query = "SELECT id,date,solvedate,taskstart.first_task_begin,task.last_task_end 
-             FROM glpi_tickets as ticket
-             INNER JOIN (
-                 select tickets_id,CASE 
-                     WHEN max(end)>max(ADDDATE(date,INTERVAL actiontime SECOND)) THEN max(end)
-                     ELSE max(ADDDATE(date,INTERVAL actiontime SECOND))
-                     END as last_task_end
-                 from glpi_tickettasks
-                 group by tickets_id) as task
-             ON ticket.id=task.tickets_id
-             LEFT JOIN (
-                 SELECT tickets_id,min(begin)AS first_task_begin
-                 FROM glpi_tickettasks
-                WHERE begin IS NOT NULL
-                GROUP BY tickets_id) AS taskstart
-             ON ticket.id=taskstart.tickets_id
-             WHERE ticket.status=5
-             AND DATE_FORMAT(ticket.solvedate, '%Y-%m-%d %H:%i')<>DATE_FORMAT(task.last_task_end, '%Y-%m-%d %H:%i')
-             AND ticket.is_deleted=0
-         AND ticket.date<task.last_task_end" . $limit;
+                FROM glpi_tickets AS ticket
+                INNER JOIN (
+                    SELECT tickets_id,CASE 
+                        WHEN max(end)>max(ADDDATE(date,INTERVAL actiontime SECOND)) THEN max(end)
+                        ELSE max(ADDDATE(date,INTERVAL actiontime SECOND))
+                        END AS last_task_end
+                    FROM glpi_tickettasks
+                    GROUP BY tickets_id) AS task
+                ON ticket.id=task.tickets_id
+                LEFT JOIN (
+                    SELECT tickets_id,min(begin)AS first_task_begin
+                    FROM glpi_tickettasks
+                    WHERE begin IS NOT NULL
+                    GROUP BY tickets_id) AS taskstart
+                ON ticket.id=taskstart.tickets_id
+                WHERE ticket.status=5
+                AND DATE_FORMAT(ticket.solvedate, '%Y-%m-%d %H:%i')<>DATE_FORMAT(task.last_task_end, '%Y-%m-%d %H:%i')
+                AND ticket.is_deleted=0
+            AND ticket.date<task.last_task_end {$limit}";
 
             foreach ($DB->doQuery($query) as $id => $row) {
                 if (!is_null($row["first_task_begin"])) {
@@ -162,11 +114,12 @@ class PluginYagpTicketsolveddate extends CommonDBTM
                 }
                 $ticket->update(['id' => $row["id"], 'solvedate' => $row["last_task_end"]]);
                 $task->addVolume(1);
-                $task->log("<a href='" . Ticket::getFormURLWithID($row["id"]) . "'>Updated Ticket id: " . $row["id"] . "</a>");
+                $tot++;
+                $href = Ticket::getFormURLWithID($row["id"]);
+                $task->log("<a href='" . $href . "'>Updated Ticket id: " . $row["id"] . "</a>");
             }
-            return true;
-        } else {
-            return false;
         }
+
+        return ($tot > 0 ? 1 : 0);
     }
 }